refactor to use helper function to fetch and draw d3 line graph
